Add files via upload
This research project examines the application of ML models in predicting the early-stage risk of diabetes and assessing the consequences of preventive measures in the health outcomes. Employing supervised learning techniques, the various symptoms and the specific demographics of a patient provided reliable predictions of risks for diabetes. This research applied Random Forest and Decision Tree classifiers for test of their effectiveness in being able to predict diabetes risk and the result was they showed high accuracy level. Research findings are consistent with what we already know by showing the necessity to take individual risk assessment into account and incorporating ML algorithms in healthcare facilities. Nevertheless, the study underlines the need for the future work in the following areas: improving model interpretability and increasing generality over the populations with large diversity. Future investigation needs extension to the method of extending the accessibility of the ML-based prediction models, checking the actual effect of them in population health and health cost in the long term, and growing the hybrid models by integrating the traditional with the current ML techniques. This research is going to find its place among much other resource-related research on ML applied analytics and clinic population management, illustrating a mean for global care provision enhancement.